requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.2
django-cors-headers==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mysqlclient==2.2.0
openai==0.27.8
Pillow==10.0.0
pytz==2023.3
redis==5.0.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
